pyproject.toml
[project]
name = "core-fraud-detection-service"
version = "1.0.0"
description = "Fraud Detection Service for Capitec SWE Assessment"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Cyril Mandla Nxumalo" }
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",

    # Cache
    "redis>=5.0.0",

    # Message Queue
    "aiokafka>=0.10.0",

    # Async Tasks (RabbitMQ broker, Redis result backend)
    "celery[librabbitmq]>=5.3.0",

    # HTTP Client
    "httpx>=0.26.0",

    # gRPC
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",

    # Rules Engine
    "pylitmus>=1.2.0",

    # Logging
    "structlog>=24.1.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",

    # Rate Limiting
    "slowapi>=0.1.9",

    # Utilities
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.1",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "fakeredis[aioredis]>=2.21.0",

    # Code Quality
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",

    # Database Testing
    "testcontainers[postgresql,redis,kafka]>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["app/grpc/generated/", "alembic/", "scripts/"]
[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function calls in defaults (needed for FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
